requirements.txt
aiogram==2.2
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
certifi==2024.6.2
channels==4.1.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
decorator==4.4.2
Django==5.0.6
django-async-orm==0.1.12
django-cors-headers==4.3.1
django-crispy-forms==2.2
django-jazzmin==3.0.0
django-unfold==0.28.1
djangorestframework==3.15.2
environs==11.0.0
frozenlist==1.4.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
marshmallow==3.21.3
moviepy==1.0.3
multidict==6.0.5
numpy==2.0.0
packaging==24.1
pillow==10.3.0
proglog==0.1.10
psycopg2-binary==2.9.9
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.1.1
sqlparse==0.5.0
TgCrypto==1.2.5
tqdm==4.66.4
urllib3==2.2.2
yarl==1.9.4
